Fix a potential score overflow.
This isn't very likely to happen. You have to pick up, like, a thousand
cell packs in the same tic to cause this. I actually do this in testing,
so let's fix it anyway.

+++ pk7/lscripts/C/src/lith_decorate.c
@@ -100,7 +100,7 @@ void Lith_UpdateScore()
 {
    player_t *p = &players[ACS_PlayerNumber()];
    
-   float rmul = RandomFloat(1.0f, 6.0f);
+   double rmul = RandomFloat(1.0f, 6.0f);
    score_t score = ACS_CheckInventory("Lith_ScoreCount") * rmul;
    
    if(ACS_GetCVar("lith_player_scoresound"))
